Refactored arePortsAvailable to Utils and added support for passing in single port. GH #5674
